requirements.txt
pyelftools>=0.24
pefile
cffi
idalink
archinfo>=7.7.9.14
pyvex>=7.7.9.14
